Move recaptcha script to head element, move BE call to assess token to recaptcha file
