+++ layouts/contact/single.html
@@ -101,5 +101,5 @@
 {{ end }}
 
 {{ define "appcode" }}
-<script src="{{ `js/contact.js` | absURL }}"></script>
+<script type="module" src="{{ `js/contact.js` | absURL }}"></script>
 {{ end }}
