+++ static/js/contact.js
@@ -1,4 +1,4 @@
-import { getRecaptchaToken } from "./reCAPTCHA";
+import { fetchAssessment } from "./recaptchaAssessment.js";
 
 // Contact Form submission
 const form = document.getElementById('contactForm');
@@ -7,23 +7,28 @@ const formID = document.getElementById('formID');
 form?.addEventListener('submit', (event) => {
   event.preventDefault();
 
-  // Get reCAPTCHA token and verify the assessment score.
+  // Get reCAPTCHA token and obtain assessment score.
   const contactBttn = document.getElementById('contact-bttn');
   const siteKey = contactBttn.dataset.sitekey;
   const action = contactBttn.dataset.action;
-
+  
   const assessment = fetchAssessment(siteKey, action);
 
+  // TODO: Display error message in the form.
   if (!assessment) {
     console.error('Unable to submit form');
     return;
   }
 
+  // If the risk analysis score is less than 0.5, do not submit the form.
+  // There is a high probability that the request is spam.
+  // TODO: Display error message in the form.
   if (assessment?.riskAnalysis?.score < 0.5) {
     console.error('Unable to submit form');
     return;
   }
 
+  // If the assessment score is greater than 0.5, submit the form.
   const formData = new FormData(form);
   const data = Object.fromEntries(formData);
 
@@ -66,31 +71,3 @@ form?.addEventListener('submit', (event) => {
       console.error('Error:', error);
     });
 });
-
-
-function fetchAssessment(sitekey, action) {
-  const token = getRecaptchaToken(sitekey, action);
-
-  if (!token) {
-    console.error('invalid recaptcha token');
-    return;
-  }
-
-  const req = {}
-  req.token = token;
-  req.action = action;
-
-  fetch(`https://api.rotationallabs.com/v1/recaptcha`, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify(req),
-  })
-    .then((response) => {
-      return response.json();
-    })
-    .catch((error) => {
-      console.error('Error:', error);
-    });
-}
