--- static/js/reCAPTCHA.js
@@ -1,6 +1,0 @@
-export function getRecaptchaToken(key, action) {
-  grecaptcha.enterprise.ready(async () => {
-    const token = await grecaptcha.enterprise.execute(key, { action: action });
-    return token;
-  })
-}
